third_party/WebKit/LayoutTests/compositing/overflow/avoid-ancestor-clip-for-scroll-children-expected.txt

Verifies that scroll children do not create ancestor clipping layers in compositor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS scrollChildHasNoScrollClip(documentLayerTree) is true
